requirements.txt
google-adk
psycopg2-binary
fastapi
uvicorn